Week 2: Entities,Repositories and Postgres BD

+++ src/main/java/ru/skypro/homework/dto/AdDTO.java
@@ -5,11 +5,11 @@ import lombok.Data;
 @Data
 public class AdDTO {
 
-    private Integer author;
+    private Long author;
 
     private String image;
 
-    private Integer pk;
+    private Long pk;
 
     private Integer price;
