+++ src/main/java/ru/skypro/homework/dto/CommentDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class CommentDTO {
-    private Integer author;
+    private Long author;
 
     private String authorImage;
 
@@ -12,7 +12,7 @@ public class CommentDTO {
 
     private Integer createdAt;
 
-    private Integer pk;
+    private Long pk;
 
     private String text;
